[REFACTORING] Rename model package to core
