+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/core/Id.scala
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance   *
  * with the License.  You may obtain a copy of the License at   *
  *                                                              *
- *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ * http://www.apache.org/licenses/LICENSE-2.0                   *
  *                                                              *
  * Unless required by applicable law or agreed to in writing,   *
  * software distributed under the License is distributed on an  *
@@ -17,11 +17,23 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.model
+package org.apache.james.jmap.core
 
+import eu.timepit.refined
 import eu.timepit.refined.api.Refined
-import eu.timepit.refined.numeric.Interval.Closed
+import eu.timepit.refined.boolean.And
+import eu.timepit.refined.collection.Size
+import eu.timepit.refined.numeric.Interval
+import eu.timepit.refined.string.MatchesRegex
 
-object StatusCode {
-  type ErrorStatus = Int Refined Closed[100, 599]
+object Id {
+  type IdConstraint = And[
+    Size[Interval.Closed[1, 255]],
+    MatchesRegex["^[#a-zA-Z0-9-_]*$"]]
+  type Id = String Refined IdConstraint
+
+  def validate(string: String): Either[IllegalArgumentException, Id] =
+    refined.refineV[IdConstraint](string)
+      .left
+      .map(new IllegalArgumentException(_))
 }
