+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/core/ProblemDetails.scala
@@ -16,22 +16,23 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
+package org.apache.james.jmap.core
 
-package org.apache.james.jmap.model
+import eu.timepit.refined.auto._
+import org.apache.http.HttpStatus.SC_BAD_REQUEST
+import org.apache.james.jmap.core.RequestLevelErrorType.ErrorTypeIdentifier
+import org.apache.james.jmap.core.StatusCode.ErrorStatus
 
-import java.time.{ZoneId, ZonedDateTime}
-import java.util.Date
+/**
+ * Problem Details for HTTP APIs within the JMAP context
+ * https://tools.ietf.org/html/rfc7807
+ * see https://jmap.io/spec-core.html#errors
+ */
+case class ProblemDetails(`type`: ErrorTypeIdentifier, status: ErrorStatus, limit: Option[String], detail: String)
 
-import org.apache.james.jmap.model.UTCDate.UTC_ZONE_ID
-
-object UTCDate {
-  private val UTC_ZONE_ID: ZoneId = ZoneId.of("UTC")
-
-  def from(date: Date, zoneId: ZoneId): UTCDate = UTCDate(ZonedDateTime.ofInstant(date.toInstant, zoneId))
-}
-
-case class UTCDate(date: ZonedDateTime) {
-  def asUTC: ZonedDateTime = {
-    date.withZoneSameInstant(UTC_ZONE_ID)
-  }
+object ProblemDetails {
+  def notRequestProblem(message: String): ProblemDetails = ProblemDetails(RequestLevelErrorType.NOT_REQUEST, SC_BAD_REQUEST, None, message)
+  def notJSONProblem(message: String): ProblemDetails = ProblemDetails(RequestLevelErrorType.NOT_JSON, SC_BAD_REQUEST, None, message)
+  def unknownCapabilityProblem(message: String): ProblemDetails = ProblemDetails(RequestLevelErrorType.UNKNOWN_CAPABILITY, SC_BAD_REQUEST, None, message)
+  def invalidResultReference(message: String): ProblemDetails = ProblemDetails(RequestLevelErrorType.UNKNOWN_CAPABILITY, SC_BAD_REQUEST, None, message)
 }
