+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/core/RequestLevelErrorType.scala
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance   *
  * with the License.  You may obtain a copy of the License at   *
  *                                                              *
- * http://www.apache.org/licenses/LICENSE-2.0                   *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
  *                                                              *
  * Unless required by applicable law or agreed to in writing,   *
  * software distributed under the License is distributed on an  *
@@ -16,28 +16,16 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
+package org.apache.james.jmap.core
 
-package org.apache.james.jmap.model
-
-import eu.timepit.refined
 import eu.timepit.refined.api.Refined
-import eu.timepit.refined.numeric.Interval.Closed
+import eu.timepit.refined.auto._
+import eu.timepit.refined.string.Uri
 
-object UnsignedInt {
-  //Unsigned int between [0, 2^53]
-  type UnsignedIntConstraint = Closed[0L, 9007199254740992L]
-  type UnsignedInt = Long Refined UnsignedIntConstraint
-
-  def validate(value: Long): Either[NumberFormatException, UnsignedInt] =
-    refined.refineV[UnsignedIntConstraint](value) match {
-      case Right(value) => Right(value)
-      case Left(error) => Left(new NumberFormatException(error))
-    }
-
-  def liftOrThrow(value: Long): UnsignedInt =
-    validate(value) match {
-      case Right(value) => value
-      case Left(error) => throw error
-    }
-
+object RequestLevelErrorType {
+  type ErrorTypeIdentifier = String Refined Uri
+  val UNKNOWN_CAPABILITY: ErrorTypeIdentifier = "urn:ietf:params:jmap:error:unknownCapability"
+  val NOT_JSON: ErrorTypeIdentifier = "urn:ietf:params:jmap:error:notJSON"
+  val NOT_REQUEST: ErrorTypeIdentifier = "urn:ietf:params:jmap:error:notRequest"
+  val LIMIT: ErrorTypeIdentifier = "urn:ietf:params:jmap:error:limit"
 }
