+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/core/ResponseObject.scala
@@ -17,23 +17,13 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.model
+package org.apache.james.jmap.core
 
-import eu.timepit.refined
-import eu.timepit.refined.api.Refined
-import eu.timepit.refined.boolean.And
-import eu.timepit.refined.collection.Size
-import eu.timepit.refined.numeric.Interval
-import eu.timepit.refined.string.MatchesRegex
+import eu.timepit.refined.auto._
+import org.apache.james.jmap.core.State.State
 
-object Id {
-  type IdConstraint = And[
-    Size[Interval.Closed[1, 255]],
-    MatchesRegex["^[#a-zA-Z0-9-_]*$"]]
-  type Id = String Refined IdConstraint
+case class ResponseObject(sessionState: State, methodResponses: Seq[Invocation])
 
-  def validate(string: String): Either[IllegalArgumentException, Id] =
-    refined.refineV[IdConstraint](string)
-      .left
-      .map(new IllegalArgumentException(_))
+object ResponseObject {
+  val SESSION_STATE: State = "75128aab4b1b"
 }
