+++ server/protocols/jmap-rfc-8621/src/test/scala/org/apache/james/jmap/core/MailboxValidationTest.scala
@@ -17,13 +17,13 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.model
+package org.apache.james.jmap.core
 
 import eu.timepit.refined.auto._
 import org.apache.james.core.Username
+import org.apache.james.jmap.core.UnsignedInt.UnsignedInt
 import org.apache.james.jmap.mail.MailboxName.MailboxName
 import org.apache.james.jmap.mail._
-import org.apache.james.jmap.model.UnsignedInt.UnsignedInt
 import org.apache.james.mailbox.model.MailboxPath
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
